modularize cards function from index.js
